feat!: allow to edit and entry
Allows the user to edit an entry in telescope. To edit an entry, you
have to open telescope, use the edit mapping (default in insert mode),
edit the new buffer however you want, then close the buffer (:q) to save
the changes

Breaking change: The edit command in telescope has the default binding
to "<C-e>" in insert mode and "e" in normal mode. This will override any
custom command with the same bindins.